app/requirements.txt
black==21.6b0
Django==3.2.4
dj-database-url==0.5.0
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
isort==5.9.1
psycopg2-binary==2.9.1
pytest-cov==2.12.1
pytest-django==4.4.0
pytest==6.2.4
whitenoise==5.2.0
